Add macro for generating owned ProcessImage structs
In some use cases it is more convenient for the wrapper struct to own
the process image buffer.  Add a second macro which generates structs
for such owned usage patterns.

Unfortunately, we still need to split mutable and immutable accessors
into separate structs as we can't be generic over mutability...