Add crate feature to allow unaligned tag addresses
By default, tags must be aligned to the size of their type (word=2,
double word=4, long word=8).  However, there are situations where data
is placed at unaligned addresses in a process image.

To make `process-image` useful for such use cases, add a crate-feature
to globally disable the assertions.  In the future, we might want to
drop the assertions entirely as their benefit is really neglible
anyway...